Commit 1.3b getting there.. slowly.
Still trying different ways to get the JTextArea to update properly. It now works with an "active total" to ensure that when another of the same item is added, it shows the correct amount and correct price. Now to just figure out how to put it all into an array and read an array into the JTextArea