Commit 1.5 - Big Update - Transaction Complete
Finished the entire transactions for card and cash and the receipt printing
Did a little visual sprucing to make it look a little nicer
Edited the initial splash screen
Admin area starting to work. Had to manipulate a JTable by placing an ArrayList<> into a DefaultTableModel first and then nesting that DTM into the JTable

Stock and Admin load correctly into the columns and rows. Now just to finalise the buttons on manipulating the data to add, edit and delete.