Commit 1.6 First Admin portal edits
Learnt that you can nest a buttons actionlistener inside another buttons action listener so that I can call upon the active tableModel inside the JTable. I have completed the add function which allows the addition of a new empty row. Will edit the layout so that the Apply Changes function only becomes visible after hitting add or delete so it prompts the user to apply the changes.